Move curl function to utility dir
